requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
aiomysql==0.2.0
Pillow==10.4.0
pyzbar==0.1.9
